prepend ExampleGroup instrumentation to the eigenclass of RSpec::Core::ExampleGroup instead of using alias_method
