+++ lib/datadog/ci/contrib/rspec/patcher.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 require "datadog/tracing/contrib/patcher"
+
 require_relative "example"
+require_relative "example_group"
 require_relative "runner"
 
 module Datadog
@@ -21,26 +23,7 @@ module Datadog
           def patch
             ::RSpec::Core::Example.include(Example)
             ::RSpec::Core::Runner.include(Runner)
-
-            ::RSpec::Core::ExampleGroup.class_eval do
-              class << self
-                alias_method :__run, :run
-
-                def run(reporter = ::RSpec::Core::NullReporter)
-                  return __run(reporter) unless top_level?
-
-                  test_suite = Datadog::CI.start_test_suite(file_path)
-                  result = __run(reporter)
-                  if result
-                    test_suite.passed!
-                  else
-                    test_suite.failed!
-                  end
-                  test_suite.finish
-                  result
-                end
-              end
-            end
+            ::RSpec::Core::ExampleGroup.include(ExampleGroup)
           end
         end
       end
